Introduce etcd's raft to implement the KV store (#222)
Currently, the controller must depend on external services like
zookeeper/etcd to store its data. And it's not so friendly for users
since they need to set up another cluster before running the controller
server. To simplify this process, we would like to introduce the raft
engine to save and replicate the data.